+++ src/main/java/com/promo/entity/ItemPromotion.java
@@ -8,7 +8,7 @@ public class ItemPromotion {
     private String sku;
     private int skuPrice;
     private int quantity;
-    private int totalPrice;
+    private float totalPrice;
     private Promotion promotion;
 
 }
